requirements.txt
fastapi
pydantic
uvicorn
python-jose
sqlalchemy
python-multipart
passlib
bcrypt